rename and address review comments
