+++ src/sentry/objectstore/__init__.py
@@ -1,0 +1,3 @@
+from sentry.objectstore.service import ObjectStoreService
+
+attachments = ObjectStoreService("attachments")
