+++ src/sentry/replays/lib/storage.py
@@ -18,8 +18,8 @@ from google.api_core.exceptions import TooManyRequests
 from sentry import options
 from sentry.models.files.file import File
 from sentry.models.files.utils import get_storage
+from sentry.objectstore.metrics import measure_storage_put
 from sentry.replays.models import ReplayRecordingSegment
-from sentry.storage.metrics import measure_storage_put
 from sentry.utils import metrics
 
 logger = logging.getLogger()
