--- src/sentry/storage/__init__.py
@@ -1,3 +1,0 @@
-from sentry.storage.service import StorageService
-
-attachments = StorageService("attachments")
